requirements.txt
# LangGraph + LangChain
langgraph>=0.3.0
langgraph-checkpoint-sqlite>=2.0.0

# LLM API
openai>=1.50.0

# RAG
rank-bm25>=0.2.2
faiss-cpu>=1.9.0
sentence-transformers>=3.0.0

# PDF 解析
pymupdf>=1.24.0

# MCP
mcp>=1.7.0

# Web
streamlit>=1.40.0

# 工具
loguru>=0.7.0
python-dotenv>=1.0.0
numpy>=1.26.0
httpx>=0.27.0

# Web Search
duckduckgo-search>=7.0.0

# 学术 API
arxiv>=2.1.0
